replace gofish-python attempt with gofish-python-2, which actually works
