--- packages/gofish-python-2/gofish/__init__.py
@@ -1,41 +1,0 @@
-"""
-GoFish Python 2 - Clean AST implementation for GoFish graphics library.
-"""
-
-from .ast import (
-    chart,
-    ChartBuilder,
-    spread,
-    stack,
-    derive,
-    group,
-    scatter,
-    rect,
-    circle,
-    line,
-    area,
-    scaffold,
-)
-
-__all__ = [
-    "chart",
-    "ChartBuilder",
-    "spread",
-    "stack",
-    "derive",
-    "group",
-    "scatter",
-    "rect",
-    "circle",
-    "line",
-    "area",
-    "scaffold",
-]
-
-__version__ = "0.1.0"
-
-
-
-
-
-
