--- packages/gofish-python-2/gofish_python_2.egg-info/SOURCES.txt
@@ -1,12 +1,0 @@
-README.md
-pyproject.toml
-gofish/__init__.py
-gofish/arrow_utils.py
-gofish/ast.py
-gofish/widget.py
-gofish_python_2.egg-info/PKG-INFO
-gofish_python_2.egg-info/SOURCES.txt
-gofish_python_2.egg-info/dependency_links.txt
-gofish_python_2.egg-info/requires.txt
-gofish_python_2.egg-info/top_level.txt
-tests/test_ast.py
