--- packages/gofish-python-2/pyproject.toml
@@ -1,32 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "gofish-python-2"
-version = "0.1.0"
-description = "Clean AST implementation for GoFish graphics library"
-readme = "README.md"
-requires-python = ">=3.8"
-dependencies = ["anywidget>=0.9.0", "pyarrow>=10.0.0", "pandas>=1.5.0"]
-# Note: The widget bundle (gofish/_static/widget.esm.js) is self-contained
-# and includes all dependencies (gofish-graphics, solid-js, apache-arrow).
-# No network access or import maps are required at runtime.
-
-[project.optional-dependencies]
-test = ["pytest>=7.0.0"]
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-python_files = "test_*.py"
-python_classes = "Test*"
-python_functions = "test_*"
-
-[tool.setuptools]
-# Include the widget bundle in package data
-# This ensures the bundle is included when building wheels
-include-package-data = true
-
-[tool.setuptools.package-data]
-# Include the widget bundle from gofish/_static/
-gofish = ["_static/widget.esm.js"]
