--- packages/gofish-python/.envrc
@@ -1,4 +1,0 @@
-# direnv configuration for uv environment
-source .venv/bin/activate
-
-
