+++ packages/gofish-python/gofish/__init__.py
@@ -1,10 +1,21 @@
 """
-GoFish Python - A Python wrapper for GoFish graphics library.
+GoFish Python 2 - Clean AST implementation for GoFish graphics library.
 """
 
-from .chart import chart, ChartBuilder
-from .operators import spread, stack, derive, group, scatter
-from .marks import rect, circle, line, area, scaffold
+from .ast import (
+    chart,
+    ChartBuilder,
+    spread,
+    stack,
+    derive,
+    group,
+    scatter,
+    rect,
+    circle,
+    line,
+    area,
+    scaffold,
+)
 
 __all__ = [
     "chart",
@@ -21,6 +32,10 @@ __all__ = [
     "scaffold",
 ]
 
-__version__ = "0.0.1"
+__version__ = "0.1.0"
+
+
+
+
