+++ packages/gofish-python/gofish/__pycache__/__init__.cpython-310.pyc
