--- packages/gofish-python/gofish/js/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "gofish-python-js-bridge",
-  "version": "0.0.1",
-  "description": "Node.js bridge for GoFish Python wrapper",
-  "type": "module",
-  "main": "render.js",
-  "scripts": {
-    "build:client": "vite build"
-  },
-  "dependencies": {
-    "apache-arrow": "^14.0.0",
-    "gofish-graphics": "file:../../../gofish-graphics",
-    "jsdom": "^23.0.0",
-    "solid-js": "^1.9.5"
-  },
-  "devDependencies": {
-    "vite": "^5.0.0",
-    "vite-plugin-solid": "^2.8.0"
-  }
-}
-
