+++ packages/gofish-python/gofish_python.egg-info/PKG-INFO
@@ -1,3 +1,15 @@
+Metadata-Version: 2.4
+Name: gofish-python
+Version: 0.1.0
+Summary: Clean AST implementation for GoFish graphics library
+Requires-Python: >=3.8
+Description-Content-Type: text/markdown
+Requires-Dist: anywidget>=0.9.0
+Requires-Dist: pyarrow>=10.0.0
+Requires-Dist: pandas>=1.5.0
+Provides-Extra: test
+Requires-Dist: pytest>=7.0.0; extra == "test"
+
 # gofish-python-2
 
 A clean, minimal implementation of the AST portion for GoFish graphics library. This package focuses solely on converting Python syntax to a JSON Intermediate Representation (IR).
