+++ packages/gofish-python/gofish_python.egg-info/SOURCES.txt
@@ -1,0 +1,13 @@
+README.md
+pyproject.toml
+gofish/__init__.py
+gofish/arrow_utils.py
+gofish/ast.py
+gofish/widget.py
+gofish/_static/widget.esm.js
+gofish_python.egg-info/PKG-INFO
+gofish_python.egg-info/SOURCES.txt
+gofish_python.egg-info/dependency_links.txt
+gofish_python.egg-info/requires.txt
+gofish_python.egg-info/top_level.txt
+tests/test_ast.py
