+++ packages/gofish-python/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gofish-python-2",
+  "name": "gofish-python",
   "version": "0.1.0",
   "type": "module",
   "private": true,
