+++ packages/gofish-python/pyproject.toml
@@ -4,22 +4,31 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "gofish-python"
-version = "0.0.1"
-description = "Python wrapper for GoFish graphics library"
+version = "0.1.0"
+description = "Clean AST implementation for GoFish graphics library"
 readme = "README.md"
 requires-python = ">=3.8"
-dependencies = [
-    "pyarrow>=10.0.0",
-    "pandas>=1.5.0",
-    "anywidget>=0.9.0",
-]
+dependencies = ["anywidget>=0.9.0", "pyarrow>=10.0.0", "pandas>=1.5.0"]
+# Note: The widget bundle (gofish/_static/widget.esm.js) is self-contained
+# and includes all dependencies (gofish-graphics, solid-js, apache-arrow).
+# No network access or import maps are required at runtime.
+
 [project.optional-dependencies]
-jupyter = ["ipython>=8.0.0"]
-all = ["ipython>=8.0.0"]
+test = ["pytest>=7.0.0"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = "test_*.py"
+python_classes = "Test*"
+python_functions = "test_*"
 
 [tool.setuptools]
+# Include the widget bundle in package data
+# This ensures the bundle is included when building wheels
+include-package-data = true
+# Explicitly specify packages to avoid including node_modules
 packages = ["gofish"]
 
 [tool.setuptools.package-data]
-gofish = ["js/dist/*.js", "js/package.json"]
-
+# Include the widget bundle from gofish/_static/
+gofish = ["_static/widget.esm.js"]
